Create channels service and small fix for loading chain on app start
