+++ packages/backend/package.json
@@ -82,7 +82,6 @@
     "mock-fs": "^5.1.2",
     "pvutils": "^1.1.3",
     "tmp": "^0.2.1",
-    "pvutils": "^1.1.3",
     "ts-jest": "^29.0.3",
     "ts-loader": "9.4.2",
     "ts-node": "10.9.1",
